Simple POST Servlet Example
Main page that is HTML that calls a servlet via a POST from a form, that
then creates a simple Person Bean (POJO) and then forwards to a simple
JSP to display the entered data. The JSP page has a button that takes
you back to the main start page.